feat: refactor form replace checkbox component and put it in form component
