+++ src/components/Form/Form.stories.tsx
@@ -1,6 +1,8 @@
+import React from "react";
 import { Meta, StoryFn } from "@storybook/react";
 import Form from "./Form";
 import { FormValues } from "./Form";
+
 export default {
   title: "Components/Form",
   component: Form,
@@ -11,5 +13,11 @@ export const Default: StoryFn = () => (
     submit={(data: FormValues) =>
       console.log("Form submitted with data:", data)
     }
-  />
+  >
+    <Form.CheckboxInput
+      name="agreeToTerms"
+      label="I agree to receive automated calls or text messages from Expert Event Solutions LLC in response to my inquiry. Msg & data rates may apply."
+      error={{ message: "You must agree to the terms" }}
+    />
+  </Form>
 );
